Refactor: Extract utility functions into separate modules
Extracted various utility functions and modal management logic from src/index.js into new, dedicated modules: src/imageUtils.js, src/modalManager.js, and src/utils.js. This improves code organization, modularity, and maintainability by separating concerns. Updated all references and imports in src/index.js accordingly.
